requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2==2.9.9
python-dotenv==0.19.0
jinja2==3.0.1
requests==2.31.0

